refactor(43-17): lift vote state to view-shell for cross-view persistence
Vote overrides previously lived locally inside WorkflowTemplateGrid and
WorkflowTemplateList. Toggling Cards <-> List unmounted the active
component and lost the optimistic vote state. Lift the state to
_view-shell.tsx via a new useVoteOverrides hook so:

- Optimistic votes persist when the user toggles between Cards and List.
- The Top rated sort uses the effective score (override ?? server),
  so a fresh upvote re-orders the cards immediately.
- The Most used sort gains a score tiebreaker so within equal-duplicate
  groups, +1 ranks above 0 ranks above -1.
