+++ components/hub/hub-results.tsx
@@ -2,10 +2,12 @@
 
 import { Search, Workflow } from "lucide-react";
 import Link from "next/link";
+import type { VoteOverridesMap } from "@/components/hub/use-vote-overrides";
 import { WorkflowTemplateGrid } from "@/components/hub/workflow-template-grid";
 import { WorkflowTemplateList } from "@/components/hub/workflow-template-list";
 import { Button } from "@/components/ui/button";
 import type { SavedWorkflow } from "@/lib/api-client";
+import type { VoteDirection } from "@/lib/workflow/editor/votes";
 
 type HubResultsProps = {
   communityWorkflows: SavedWorkflow[];
@@ -14,6 +16,8 @@ type HubResultsProps = {
   featuredIds?: Set<string>;
   onClearFilters?: () => void;
   viewMode: "cards" | "list"; // HUB-18 — branches on cookie-driven view
+  voteOverrides: VoteOverridesMap;
+  onVote: (workflowId: string, direction: VoteDirection) => Promise<void>;
 };
 
 export function HubResults({
@@ -23,6 +27,8 @@ export function HubResults({
   featuredIds,
   onClearFilters,
   viewMode,
+  voteOverrides,
+  onVote,
 }: HubResultsProps): React.ReactElement {
   const workflows = isSearchActive ? searchResults : communityWorkflows;
 
@@ -63,14 +69,24 @@ export function HubResults({
   if (viewMode === "list") {
     return (
       <section data-view-mode="list">
-        <WorkflowTemplateList featuredIds={featuredIds} workflows={workflows} />
+        <WorkflowTemplateList
+          featuredIds={featuredIds}
+          onVote={onVote}
+          voteOverrides={voteOverrides}
+          workflows={workflows}
+        />
       </section>
     );
   }
 
   return (
     <section data-view-mode="cards">
-      <WorkflowTemplateGrid featuredIds={featuredIds} workflows={workflows} />
+      <WorkflowTemplateGrid
+        featuredIds={featuredIds}
+        onVote={onVote}
+        voteOverrides={voteOverrides}
+        workflows={workflows}
+      />
     </section>
   );
 }
